move command module
