--- docs/api/utils/command.rst
@@ -1,5 +1,0 @@
-qcodes.utils.command
---------------------
-
-.. automodule:: qcodes.utils.command
-   :members:
