+++ qcodes/parameters/function.py
@@ -1,8 +1,9 @@
 from typing import TYPE_CHECKING, Any, Callable, List, Optional, Sequence, Union
 
-from qcodes.utils.command import Command
 from qcodes.utils.metadata import Metadatable
 from qcodes.utils.validators import Validator, validate_all
+
+from .command import Command
 
 if TYPE_CHECKING:
     from qcodes.instrument import InstrumentBase
